refactor(#792): decompose block_parameter_wiring::wire (727-line fn) by responsibility
Orchestrator + 4 helpers by param type: wire_numeric_params, wire_text_bool_params (base)
+ wire_select_param, wire_toggle_and_file (block_parameter_extras.rs sibling). Same
closures/captures via &BlockParameterCtx. File 794->~460, off debt. block_parameter tests: 4 pass, zero warnings.
