+++ crates/adapter-gui/src/lib.rs
@@ -26,6 +26,7 @@ mod block_insert_callbacks;
 mod block_model_search_wiring;
 mod block_panel_dimensions;
 mod block_parameter_wiring;
+mod block_parameter_extras;
 mod block_picker_wiring;
 /// #614: compact chain view callbacks — also exposes public play/stop helpers
 /// for integration tests (`compact_chain_di_loop_play`, `compact_chain_di_loop_stop`).
